Configure the interface before bringing it up
Currently, clatd brings the clat4 interface up before assigning
its IPv4 address. This can cause a race condition, because as
soon as the interface comes up, the framework notices and reads
the interface configuration into the LinkProperties. If this
happens before the IPv4 address is configured, then the framework
ends up thinking clat4's IPv4 address is 0.0.0.0/0.

Fix this by configuring the address before the interface is
brought up.

Currently the framework does not use this address for anything,
so this is purely cosmetic, but it could make debugging more
confusing.

Bug: 8276725
Change-Id: I2bfee586a0d70050c53b10cc3f7eb9a98173e11d