Delete the clatd code that assigns IPv4 and IPv6 addresses.
The IPv4 and IPv6 address assignment code, and its tests, were
migrated to ClatdController in Q, and this code has been unused
since then.

This code is not only unused, it also *should* never be used,
because if it were, many parts of 464xlat on Android would not
work as expected. This includes 464xlat BPF offload and data
usage accounting, both of which require netd to know exactly
which IPv4 and IPv6 addresses are used by clatd at any given
time.

Most of the code is deleted, but the function to generate random
checksum-neutral IIDs was moved to the tests instead.

This change does not delete the configuration file parsing code.
That will be done in a follow-up CL.

Bug: 144730808
Bug: 150738490
Test: atest clatd_test
Test: atest netd_integration_test
Test: IPv6-only wifi continues to work
Original-Change: https://android-review.googlesource.com/1320923
Merged-In: I8265e9bb04d44ed1dda445fa91881fcd1601dc47
Change-Id: I8265e9bb04d44ed1dda445fa91881fcd1601dc47