Implement ConnectingState
This change implements the Connecting state, where the gateway
connection will be established. In the event of a network switch during
connection, the existing connection attempt will be aborted and
restarted on the new network in an attempt to avoid the case where the
old network is completely lost (eg wifi to cell switch), and IKE stalls
for the entire retry period.

Bug: 165827287
Test: atest FrameworksVcnTests
Change-Id: Id1c6337e8ba4cc1b97c6c20977153c230bd07723