Define TRANSPORT_USB in NetworkCapabilitiesUtils temporarily
The new API constant - TRANSPORT_USB didn't in mainline
branch, so the user of TRANSPORT_USB will get build break
when it is built in mainline branch.
Define TRANSPORT_USB in NetworkCapabilitiesUtils temporarily
is a workaround to fix this problem, and it should be removed
once downstream branches are S-based.

Bug: 184158327
Test: m
Change-Id: I83560c75761e7008b457a66296364cb4ce24db50
Ignore-AOSP-First: This should be compiled with ag/14049950.