Add DisableConfigSyncTargetPreparer
The TargetPreparer disables DeviceConfig sync before a test module, and
attempts to reset it to its original state afterwards.

It only supports S+ as the associated commands are not available on R-.
The format of commands has also changed in some T branches , so support
legacy format if the newer format is not supported by the device.

Bug: 210377950
Test: atest CtsNetTestCases using the preparer, verify setting set
      during the test.
Change-Id: Iaabaa1050b024e1378ab1cf87a394cc8ebab227b