Sync up the latest change in NetworkMonitor.
There are two copies of NetlinkMonitor, one resides inside NetworkStack
module and it's shared with Tethering module via filegroup, the other
one is in frameworks/libs/net. Since this class is shared among of
modules we should just maintain it in frameworks/libs/net and remove
the duplicated one from NetworkStack module. Sync up the latest change
in NetworkMonitor made in NetworkStack.

Bug: 235901424
Test: m
Change-Id: Ia221f35cfd6764a3e106873e476258cd3298a75a