Add test for partial responses
Test that when a responder only responds with the exact records that
were queried, so only reply for PTR in discovery, only send SRV, TXT,
A, AAAA when asked explicitly, service resolve succeeds.

This ensures that the querier sends followup queries for each record.
See RFC6763 12., especially the last paragraph.

Bug: 267570781
Bug: 267371243
Test: atest NsdManagerTest
Change-Id: Ia392e80c1e27b479c6177d19f6b4be6032dcb1cd