Remove unnecessary cts tag.
Some modules defined in the build file are only data dependencies of cts test modules. In this case, they will be packaged to corresponding test modules. For example, if module A is a data dependency of test module B, the module A can be found under android-cts/testcases/B/ directory. Therefore it is unnecesasry to tag data dependencies which will cause a duplicate folder, e.g. android-cts/testcases/A.

Change-Id: I1b0f08c245dc51419ed12283c40df2e6ee71ebce
Test: m cts
Bug: 283199104
Ignore-AOSP-First: The clean-up is for package size improvement on U. The backfill will apply to aosp later.